Add entity parameter on Identity generator method
To avoid that different classes with the same method
share the same cache item, the identity generator methods
should to use the entity to generate an id.

Eg.: we have two classes:

MyProject\Newsletter\Repository\User::findById($id)
MyProject\Registration\Repository\User::findById($id)

Tardis need to use the namespace for example to know
that we are caching different things.